upd ci again, reduce ci time, split into two ci files
