+++ test/registered/radix_cache/test_unified_radix_cache_kl.py
@@ -35,10 +35,7 @@ MAMBA_TRACK_INTERVAL = 128
 SWA_MODEL = "openai/gpt-oss-20b"
 FULL_MODEL = "Qwen/Qwen3-32B"
 
-DSV4_FLASH_MODEL = "sgl-project/DeepSeek-V4-Flash-FP8"
-DSV4_FLASH_LAUNCH_TIMEOUT = 3600
-
-register_cuda_ci(est_time=1200, stage="base-c", runner_config="8-gpu-h200")
+register_cuda_ci(est_time=760, stage="base-c", runner_config="4-gpu-h100")
 
 
 class UnifiedRadixTreeTestMixin:
@@ -266,146 +263,6 @@ class TestUnifiedSWARadixCache(UnifiedRadixTreeTestMixin, CustomTestCase):
                 "--disable-piecewise-cuda-graph",
             ],
             env={"SGLANG_ENABLE_UNIFIED_RADIX_TREE": "1"},
-        )
-        cls.input_ids = get_input_ids(cls.model, num_samples=18)
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-
-class TestUnifiedMambaHiCache(UnifiedRadixTreeTestMixin, CustomTestCase):
-    """Mamba hybrid + HiCache + UnifiedRadixCache."""
-
-    kl_threshold = 0.005
-    prefill_cache_assert = staticmethod(
-        make_mamba_prefill_assert(chunk_size=MAMBA_CHUNK_SIZE)
-    )
-    decode_cache_assert = staticmethod(
-        make_mamba_decode_assert(track_interval=MAMBA_TRACK_INTERVAL)
-    )
-
-    @unittest.skipIf(is_in_ci(), "To reduce the CI execution time.")
-    def test_gsm8k(self):
-        pass
-
-    @unittest.skipIf(is_in_ci(), "To reduce the CI execution time.")
-    def test_gsm8k(self):
-        pass
-
-    @classmethod
-    def setUpClass(cls):
-        cls.model = MAMBA_MODEL
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=[
-                "--tp-size",
-                "4",
-                "--chunked-prefill-size",
-                "2048",
-                "--mem-fraction-static",
-                "0.85",
-                "--mamba-scheduler-strategy",
-                "extra_buffer",
-                "--mamba-track-interval",
-                str(MAMBA_TRACK_INTERVAL),
-                "--enable-hierarchical-cache",
-                "--hicache-ratio",
-                "4",
-                "--hicache-write-policy",
-                "write_through",
-                "--hicache-io-backend",
-                "direct",
-                "--hicache-mem-layout",
-                "page_first_direct",
-                "--max-total-tokens",
-                "12000",
-                "--max-mamba-cache-size",
-                "500",
-                "--max-running-requests",
-                "4",
-            ],
-            env={"SGLANG_ENABLE_UNIFIED_RADIX_TREE": "1"},
-        )
-        cls.input_ids = get_input_ids(cls.model, num_samples=18)
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-
-def _assert_dsv4_decode_cached_tokens(result, history_len, output_len, label):
-    expected = history_len + output_len
-    actual = result["meta_info"]["cached_tokens"]
-    lower = max(0, expected - 256)
-    assert actual >= lower, f"{label}: expected cached_tokens>={lower}, got {actual}"
-
-
-class TestUnifiedDeepSeekV4FlashHiCache(UnifiedRadixTreeTestMixin, CustomTestCase):
-    """DeepSeek V4 Flash FP8 + HiCache + UnifiedRadixCache."""
-
-    kl_threshold = 0.005
-    sampling_temperature = 0
-    decode_cache_assert = staticmethod(_assert_dsv4_decode_cached_tokens)
-    gsm8k_threshold = 0.90
-    num_gsm8k_questions = 100
-
-    @unittest.skipIf(is_in_ci(), "To reduce the CI execution time.")
-    def test_gsm8k(self):
-        pass
-
-    @unittest.skipIf(is_in_ci(), "To reduce the CI execution time.")
-    def test_gsm8k(self):
-        pass
-
-    @unittest.skipIf(is_in_ci(), "To reduce the CI execution time.")
-    def test_multiturn_logprobs_match(self):
-        pass
-
-    @classmethod
-    def setUpClass(cls):
-        cls.model = DSV4_FLASH_MODEL
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DSV4_FLASH_LAUNCH_TIMEOUT,
-            other_args=[
-                "--trust-remote-code",
-                "--tp-size",
-                "4",
-                "--attention-backend",
-                "compressed",
-                "--page-size",
-                "256",
-                "--chunked-prefill-size",
-                "8192",
-                "--mem-fraction-static",
-                "0.9",
-                "--disable-shared-experts-fusion",
-                "--enable-hierarchical-cache",
-                "--hicache-ratio",
-                "4",
-                "--hicache-write-policy",
-                "write_through",
-                "--hicache-io-backend",
-                "direct",
-                "--hicache-mem-layout",
-                "page_first_direct",
-                "--swa-full-tokens-ratio",
-                "0.25",
-                "--max-total-tokens",
-                "20000",
-                "--max-running-requests",
-                "2",
-            ],
-            env={
-                "SGLANG_DSV4_FP4_EXPERTS": "0",
-                "SGLANG_ENABLE_UNIFIED_RADIX_TREE": "1",
-            },
         )
         cls.input_ids = get_input_ids(cls.model, num_samples=18)
